Revert "[SSN-180]: 로그인 플로우 수정"
